requirements.txt
fastapi
uvicorn
pymongo
selenium
webdriver_manager
jinja2
python-multipart
python-dotenv